+++ saleor/graphql/channel/mutations/channel_update.py
@@ -5,7 +5,7 @@ from django.utils.text import slugify
 from ....channel import models
 from ....channel.error_codes import ChannelErrorCode
 from ....core.tracing import traced_atomic_transaction
-from ....permission.enums import ChannelPermissions
+from ....permission.enums import ChannelPermissions, OrderPermissions
 from ....shipping.tasks import (
     drop_invalid_shipping_methods_relations_for_given_channels,
 )
@@ -53,10 +53,16 @@ class ChannelUpdate(ModelMutation):
         )
 
     class Meta:
-        description = "Update a channel."
+        description = (
+            "Update a channel.\n\n"
+            "Requires one of the following permissions: MANAGE_CHANNELS.\n"
+            "Requires one of the following permissions "
+            "when updating only orderSettings field: "
+            "MANAGE_CHANNELS, MANAGE_ORDERS."
+        )
+        auto_permission_message = False
         model = models.Channel
         object_type = Channel
-        permissions = (ChannelPermissions.MANAGE_CHANNELS,)
         error_type_class = ChannelError
         error_type_field = "channel_errors"
 
@@ -84,8 +90,29 @@ class ChannelUpdate(ModelMutation):
             cleaned_input["slug"] = slugify(slug)
         if stock_settings := cleaned_input.get("stock_settings"):
             cleaned_input["allocation_strategy"] = stock_settings["allocation_strategy"]
+        if order_settings := cleaned_input.get("order_settings"):
+            cleaned_input["automatically_confirm_all_new_orders"] = order_settings[
+                "automatically_confirm_all_new_orders"
+            ]
+            cleaned_input[
+                "automatically_fulfill_non_shippable_gift_card"
+            ] = order_settings["automatically_fulfill_non_shippable_gift_card"]
 
         return cleaned_input
+
+    @classmethod
+    def check_permissions(cls, context, permissions=None, **data):
+        permissions = [ChannelPermissions.MANAGE_CHANNELS]
+
+        # Permission MANAGE_ORDERS or MANAGE_CHANNELS is required to update
+        # orderSettings. MANAGE_ORDERS can be used only when
+        # channelUpdate mutation will get OrderSettingsInput and
+        # not any other fields. Otherwise needed permission is MANAGE_CHANNELS.
+        input = data["data"]["input"]
+        if "order_settings" in input and len(input) == 1:
+            permissions.append(OrderPermissions.MANAGE_ORDERS)
+
+        return super().check_permissions(context, permissions, **data)
 
     @classmethod
     def _save_m2m(cls, info: ResolveInfo, instance, cleaned_data):