+++ saleor/graphql/payment/mutations.py
@@ -692,7 +692,7 @@ class TransactionCreate(BaseMutation):
         permissions = (PaymentPermissions.HANDLE_PAYMENTS,)
 
     @classmethod
-    def check_permissions(cls, context, permissions=None):
+    def check_permissions(cls, context, permissions=None, **data):
         """Determine whether app has rights to perform this mutation."""
         permissions = permissions or cls._meta.permissions
         if app := getattr(context, "app", None):
@@ -1018,7 +1018,7 @@ class TransactionRequestAction(BaseMutation):
         )
 
     @classmethod
-    def check_permissions(cls, context, permissions=None):
+    def check_permissions(cls, context, permissions=None, **data):
         required_permissions = permissions or cls._meta.permissions
         requestor = get_user_or_app_from_context(context)
         for required_permission in required_permissions: